+++ frontend/src/App.tsx
@@ -1,3 +1,5 @@
+import LoginRoute from "@/routes/Login";
+import RegisterRoute from "@/routes/Register";
 import { BrowserRouter, Route, Routes } from "react-router";
 
 function App() {
@@ -5,6 +7,8 @@ function App() {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<div>Home Route</div>} />
+        <Route path="/login" element={<LoginRoute />} />
+        <Route path="/register" element={<RegisterRoute />} />
       </Routes>
     </BrowserRouter>
   );